Add favorites display
Took a lot of tweaking to realize I was making a dumb dumb dumb dumb mistake. Coding this late is not good! Cleanup will include refactoring where available, adding significant CSS changes to indicate which panel of the website is active, and somehow probably purging my server of mock logins